finish momo
hoàn thành tích hợp

+++ HeadPhone/database/migrations/2024_03_02_123721_create_order_details_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
         Schema::create('order_details', function (Blueprint $table) {
             $table->id();
             $table->string('color');
-            $table->smallInteger('quantity');
+            $table->integer('quantity');
             $table->double('price');
             $table->string('order_id')->references('id')->on('orders');
             $table->foreignIdFor(Product::class)->constrained()->cascadeOnDelete();
